加密用户qq号

+++ src/main/java/cn/zzwtsy/pu/dao/UserDao.java
@@ -2,6 +2,7 @@ package cn.zzwtsy.pu.dao;
 
 import cn.zzwtsy.pu.bean.User;
 import cn.zzwtsy.pu.database.DataBaseHelper;
+import cn.zzwtsy.pu.utils.Encryption;
 
 import java.sql.SQLException;
 import java.util.Map;
@@ -13,6 +14,8 @@ import java.util.Map;
  * @since 2022/12/23
  */
 public class UserDao {
+    //盐值
+    private static final String salt = "PuCamps";
     private final String tableName = "user";
 
     /**
@@ -22,14 +25,18 @@ public class UserDao {
      * @return {@link User}
      */
     public User getUserByQqId(long qqId) {
+        String encryption = Encryption.encryptionToMd5(String.valueOf(qqId), salt);
+        if (encryption == null || encryption.isEmpty()) {
+            return null;
+        }
         User user = new User();
         String sql = "SELECT * FROM " + tableName + " WHERE qqId = ?";
         try {
-            Map<Object, Object> queryMap = DataBaseHelper.executeQueryMap(sql, qqId);
+            Map<Object, Object> queryMap = DataBaseHelper.executeQueryMap(sql, encryption);
             if (queryMap != null) {
                 user.setOauthToken((String) queryMap.get("oauthToken"));
                 user.setOauthTokenSecret((String) queryMap.get("oauthTokenSecret"));
-                user.setQqId((long) queryMap.get("qqId"));
+                user.setQqId((String) queryMap.get("qqId"));
                 user.setUid((String) queryMap.get("uid"));
             } else {
                 return null;
@@ -48,10 +55,14 @@ public class UserDao {
      * @return 受影响的行数: >=1 添加成功
      */
     public int addUser(long qqId, String uid, String oauthToken, String oauthTokenSecret) {
+        String encryption = Encryption.encryptionToMd5(String.valueOf(qqId), salt);
+        if (encryption == null || encryption.isEmpty()) {
+            return 0;
+        }
         int status;
         String sql = "INSERT INTO " + tableName + " (qqId,uid,oauthToken,oauthTokenSecret) VALUES (?,?,?,?)";
         try {
-            status = DataBaseHelper.executeUpdate(sql, qqId, uid, oauthToken, oauthTokenSecret);
+            status = DataBaseHelper.executeUpdate(sql, encryption, uid, oauthToken, oauthTokenSecret);
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
@@ -65,10 +76,14 @@ public class UserDao {
      * @return 受影响的行数: >=1 删除成功
      */
     public int deleteUser(long qqId) {
+        String encryption = Encryption.encryptionToMd5(String.valueOf(qqId), salt);
+        if (encryption == null || encryption.isEmpty()) {
+            return 0;
+        }
         int deleteStatus;
         String sql = "DELETE FROM" + tableName + "WHERE qqId = ?";
         try {
-            deleteStatus = DataBaseHelper.executeUpdate(sql, qqId);
+            deleteStatus = DataBaseHelper.executeUpdate(sql, encryption);
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
@@ -84,10 +99,14 @@ public class UserDao {
      * @return 受影响的行数: >=1 更新成功
      */
     public int updateUser(long qqId, String uid, String oauthToken, String oauthTokenSecret) {
+        String encryption = Encryption.encryptionToMd5(String.valueOf(qqId), salt);
+        if (encryption == null || encryption.isEmpty()) {
+            return 0;
+        }
         int status;
         String sql = "UPDATE " + tableName + " SET uid = ?,oauthToken = ?,oauthTokenSecret = ? WHERE qqId = ?";
         try {
-            status = DataBaseHelper.executeUpdate(sql, uid, oauthToken, oauthTokenSecret, qqId);
+            status = DataBaseHelper.executeUpdate(sql, uid, oauthToken, oauthTokenSecret, encryption);
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
